+++ lang/cs/articles.php
@@ -1,6 +1,6 @@
 <?php
 
-return array(
+return [
     'count' => 'článek|články|článků',
     'no_results' => 'momentálně žádné články',
     'filter' => [
@@ -9,6 +9,7 @@ return array(
         'sort_by' => [
             'date_desc' => 'data',
             'date_asc' => 'data od nejstarších',
-        ]
-    ]
-);
+        ],
+    ],
+    'we_recommend' => 'Doporučujeme',
+];