+++ CHANGELOG.md
@@ -4,8 +4,9 @@ All notable changes to this project will be documented in this file[^1].
 ## [Unreleased]
 ### Fixed
 - indexing of Authority roles
+- error component attempting to show map with Item location
 
-### Added 
+### Added
 - parse article teaser from admin
 
 ## [2.68.0] - 2022-10-13